Integrate with arrival time analysis and move to datapath
