+++ lib/Dialect/Datapath/CMakeLists.txt
@@ -13,6 +13,7 @@ add_circt_dialect_library(CIRCTDatapath
   Support
 
   LINK_LIBS PUBLIC
+  CIRCTAIGAnalysis
   CIRCTHW
   CIRCTComb
   MLIRIR
